requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Login==0.6.3
gunicorn==25.0.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
psycopg2-binary
cloudinary
reportlab